Change raised error message to english
This was originally in portuguese because it was intended to be a
message that could be returned to the end user. But considering that it
turned to be an exception error message I don't think the developer (user of the
library) should rescue it and forward to the end user.

With that in mind, this message is intended to be informative to the
developer.